Port SqlClient and WindowsIdentity x-plat fixes to RC2
This is a port of the following changes:
6902316fed76f2a905ef626b7ef0a64b4749a237
Remove dependencies on WindowsIdentity from Unix binaries
It looks like a couple libraries were using
System.Security.Principal.Windows unintentionally in their
Unix implementations.  @dagood discovered this when
trying to restore our tools-runtime project for unix RIDs.

This fix removes the dependencies from those libs
in the CoreFx repro.

78977c160df8be548c4dc9ac512f45ebb3382ff8
Make sure to default to the Windows build of SqlClient
