+++ src/System.Data.SqlClient/src/System.Data.SqlClient.builds
@@ -2,7 +2,12 @@
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$([MSBuild]::GetDirectoryNameOfFileAbove($(MSBuildThisFileDirectory), dir.props))\dir.props" />
   <ItemGroup>
-    <Project Include="System.Data.SqlClient.csproj" />
+    <Project Include="System.Data.SqlClient.csproj">
+      <OSGroup>Linux</OSGroup>
+    </Project>
+    <Project Include="System.Data.SqlClient.csproj">
+      <OSGroup>Windows_NT</OSGroup>
+    </Project>
     <Project Include="facade\System.Data.SqlClient.csproj">
       <TargetGroup>net46</TargetGroup>
     </Project>
