+++ src/System.Data.SqlClient/src/System/Data/ProviderBase/DbConnectionPoolIdentity.cs
@@ -9,10 +9,9 @@ using System.Security.Principal;
 
 namespace System.Data.ProviderBase
 {
-    sealed internal class DbConnectionPoolIdentity
+    sealed internal partial class DbConnectionPoolIdentity
     {
         static public readonly DbConnectionPoolIdentity NoIdentity = new DbConnectionPoolIdentity(String.Empty, false, true);
-        static private DbConnectionPoolIdentity s_lastIdentity = null;
 
         private readonly string _sidString;
         private readonly bool _isRestricted;
@@ -42,32 +41,6 @@ namespace System.Data.ProviderBase
                 result = ((_sidString == that._sidString) && (_isRestricted == that._isRestricted) && (_isNetwork == that._isNetwork));
             }
             return result;
-        }
-
-        static internal DbConnectionPoolIdentity GetCurrent()
-        {
-            DbConnectionPoolIdentity current;
-            using (WindowsIdentity identity = WindowsIdentity.GetCurrent())
-            {
-                IntPtr token = identity.AccessToken.DangerousGetHandle();
-                bool isNetwork = identity.User.IsWellKnown(WellKnownSidType.NetworkSid);
-                string sidString = identity.User.Value;
-
-                // Win32NativeMethods.IsTokenRestricted will raise exception if the native call fails
-                bool isRestricted = Win32NativeMethods.IsTokenRestrictedWrapper(token);
-
-                var lastIdentity = s_lastIdentity;
-                if ((lastIdentity != null) && (lastIdentity._sidString == sidString) && (lastIdentity._isRestricted == isRestricted) && (lastIdentity._isNetwork == isNetwork))
-                {
-                    current = lastIdentity;
-                }
-                else
-                {
-                    current = new DbConnectionPoolIdentity(sidString, isRestricted, isNetwork);
-                }
-            }
-            s_lastIdentity = current;
-            return current;
         }
 
         override public int GetHashCode()
